+++ packages/core/src/components/header/header.i18n.ts
@@ -1,0 +1,24 @@
+import { Resource } from "../../utils/i18n";
+
+export const translations: Resource = {
+  en: {
+    "dso-header": {
+      menu: "Menu",
+      login: "Log in",
+      logout: "Log out",
+      help: "Help",
+      overflowMenu: "More",
+      userHome: "Mijn Omgevingsloket",
+    },
+  },
+  nl: {
+    "dso-header": {
+      menu: "Menu",
+      login: "Inloggen",
+      logout: "Uitloggen",
+      help: "Help",
+      overflowMenu: "Meer",
+      userHome: "Mijn Omgevingsloket",
+    },
+  },
+};